+++ opentelemetry-sdk/tests/test_configurator.py
@@ -252,7 +252,7 @@ class DummyOTLPLogExporter(LogRecordExporter):
     def shutdown(self):
         pass
 
-    def force_flush(self, timeout_millis: float = 10_000) -> bool:
+    def force_flush(self, timeout_millis: int = 10_000) -> bool:
         return True
